Renamed openai-common package to just common
